LibC: Move even more methods and globals out of crt0.o
