+++ Libraries/LibC/crt0.cpp
@@ -7,18 +7,10 @@ extern "C" {
 
 int main(int, char**, char**);
 
-__thread int errno;
-char** environ;
-bool __environ_is_malloced;
-
-void __libc_init()
-{
-    void __malloc_init();
-    __malloc_init();
-
-    void __stdio_init();
-    __stdio_init();
-}
+extern void __libc_init();
+extern void _init();
+extern char** environ;
+extern bool __environ_is_malloced;
 
 int _start(int argc, char** argv, char** env)
 {
@@ -27,7 +19,6 @@ int _start(int argc, char** argv, char** env)
 
     __libc_init();
 
-    extern void _init();
     _init();
 
     extern void (*__init_array_start[])(int, char**, char**) __attribute__((visibility("hidden")));
@@ -42,18 +33,5 @@ int _start(int argc, char** argv, char** env)
     exit(status);
 
     return 20150614;
-}
-
-[[noreturn]] void __cxa_pure_virtual()
-{
-    ASSERT_NOT_REACHED();
-}
-
-extern u32 __stack_chk_guard;
-u32 __stack_chk_guard = (u32)0xc0000c13;
-
-[[noreturn]] void __stack_chk_fail()
-{
-    ASSERT_NOT_REACHED();
 }
 }
